new changes and renamed some files
