+++ src/views/forgot.jsx
