+++ src/views/register.jsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-
+import Button from '@material-ui/core/Button';
 import axios from 'axios';
 
 export default class Register extends Component{
